push 1 28/12

+++ app/Http/Controllers/LegalInformationController.php
@@ -15,6 +15,8 @@ class LegalInformationController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', legalInformation::class);
+
         $legalInformations = legalInformation::orderBy('type')->get();
         return view('LegalInformations.index', compact('legalInformations'));
     }
@@ -26,6 +28,8 @@ class LegalInformationController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', legalInformation::class);
+
         return view('LegalInformations.create');
     }
 
@@ -37,6 +41,8 @@ class LegalInformationController extends Controller
      */
     public function store(LegalInformationRequest $request)
     {
+        $this->authorize('create', legalInformation::class);
+
         $legalInformation                 = new legalInformation();
         $legalInformation->type           = $request->get('type');
         $legalInformation->description    = $request->get('description');
@@ -56,6 +62,8 @@ class LegalInformationController extends Controller
      */
     public function show(LegalInformation $legalInformation)
     {
+        $this->authorize('view', legalInformation::class);
+
         return view('LegalInformations.show', compact('legalInformation'));
     }
 
@@ -67,6 +75,8 @@ class LegalInformationController extends Controller
      */
     public function edit(LegalInformation $legalInformation)
     {
+        $this->authorize('update', legalInformation::class);
+
         return view('LegalInformations.edit', compact('legalInformation'));
     }
 
@@ -79,6 +89,8 @@ class LegalInformationController extends Controller
      */
     public function update(LegalInformationRequest $request, LegalInformation $legalInformation)
     {
+        $this->authorize('update', legalInformation::class);
+
         $legalInformation->type = $request->get('type');
         $legalInformation->description = $request->get('description');
 
@@ -97,6 +109,8 @@ class LegalInformationController extends Controller
      */
     public function destroy(LegalInformation $legalInformation)
     {
+        $this->authorize('delete', legalInformation::class);
+
         if($legalInformation->delete()){
             $message = 'Your delete processed correctly';
         }
