+++ app/Http/Controllers/ResearchOutputController.php
@@ -24,6 +24,8 @@ class ResearchOutputController extends Controller
      */
     public function index(Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project)
     {
+        $this->authorize('viewAny', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project);
+
         $researchOutputs = $project->researchOutputs()->orderBy('title')->get();
 
         return view('ResearchOutputs.index', compact('node', 'educationalInstitution', 'faculty', 'researchGroup', 'researchTeam', 'project', 'researchOutputs'));
@@ -36,8 +38,10 @@ class ResearchOutputController extends Controller
      */
     public function create(Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project)
     {
+        $this->authorize('create', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project);
+
         $mincienciasTypologies = json_decode(Storage::get('public/json/minciencias_typologies.json'), true);
-        
+
         return view('ResearchOutputs.create', compact('node', 'educationalInstitution', 'faculty', 'researchGroup', 'researchTeam', 'project', 'mincienciasTypologies'));
     }
 
@@ -49,6 +53,8 @@ class ResearchOutputController extends Controller
      */
     public function store(ResearchOutputRequest $request, Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project)
     {
+        $this->authorize('create', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project);
+
         $researchOutput = new ResearchOutput();
         $researchOutput->title          = $request->get('title');
         $researchOutput->typology       = $request->get('typology');
@@ -79,6 +85,8 @@ class ResearchOutputController extends Controller
      */
     public function show(Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project, ResearchOutput $researchOutput)
     {
+        $this->authorize('view', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project,$researchOutput);
+
         return view('ResearchOutputs.show', compact('node', 'educationalInstitution', 'faculty', 'researchGroup', 'researchTeam', 'project', 'researchOutput'));
     }
 
@@ -90,6 +98,8 @@ class ResearchOutputController extends Controller
      */
     public function edit(Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project, ResearchOutput $researchOutput)
     {
+        $this->authorize('update', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project,$researchOutput);
+
         $mincienciasTypologies = json_decode(Storage::get('public/json/minciencias_typologies.json'), true);
 
         return view('ResearchOutputs.edit', compact('node', 'educationalInstitution', 'faculty', 'researchGroup', 'researchTeam', 'project', 'researchOutput', 'mincienciasTypologies'));
@@ -104,6 +114,8 @@ class ResearchOutputController extends Controller
      */
     public function update(ResearchOutputRequest $request, Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project, ResearchOutput $researchOutput)
     {
+        $this->authorize('update', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project,$researchOutput);
+
         $researchOutput->title          = $request->get('title');
         $researchOutput->typology       = $request->get('typology');
         $researchOutput->description    = $request->get('description');
@@ -131,6 +143,8 @@ class ResearchOutputController extends Controller
      */
     public function destroy(Node $node, EducationalInstitution $educationalInstitution, EducationalInstitutionFaculty $faculty, ResearchGroup $researchGroup, ResearchTeam $researchTeam, Project $project, ResearchOutput $researchOutput)
     {
+        $this->authorize('delete', ResearchOutput::class , $node, $educationalInstitution , $faculty, $researchGroup,$researchTeam,$project,$researchOutput);
+
         if($researchOutput->delete()){
             $message = 'Your delete processed correctly';
         }
