+++ app/Http/Controllers/RoleController.php
@@ -15,6 +15,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Role::class);
+
         $roles = Role::orderBy('name')->get();
 
         return view('Roles.index', compact('roles'));
@@ -27,6 +29,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Role::class);
+
         $permissions = Permission::orderBy('model')->get();
 
         return view('Roles.create', compact('permissions'));
@@ -40,6 +44,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Role::class);
+
         $role = new Role();
         $role->name        = $request->name;
         $role->description = $request->description;
@@ -60,6 +66,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        $this->authorize('view', Role::class, $role);
+
         return view('Roles.show', compact('role'));
     }
 
@@ -71,6 +79,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('update', Role::class, $role);
+
         $permissions = Permission::orderBy('model')->get();
 
         return view('Roles.edit', compact('role', 'permissions'));
@@ -85,6 +95,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        $this->authorize('update', Role::class, $role);
+
         $role->name         = $request->name;
         $role->description  = $request->description;
         $role->guard_name   = 'web';
@@ -104,6 +116,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('delete', Role::class, $role);
+
         if($role->delete()) {
             $message = 'Your delete processed correctly';
         }
